completed custom fields with multi language

+++ apps/web-giddh/src/app/custom-fields/list/list.component.ts
@@ -1,10 +1,13 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit } from "@angular/core";
+import { MatDialog } from "@angular/material/dialog";
 import { PageChangedEvent } from "ngx-bootstrap/pagination";
 import { ReplaySubject } from "rxjs";
-import { takeUntil } from "rxjs/operators";
+import { take, takeUntil } from "rxjs/operators";
 import { PAGINATION_LIMIT } from "../../app.constant";
 import { CustomFieldsService } from "../../services/custom-fields.service";
 import { ToasterService } from "../../services/toaster.service";
+import { ConfirmModalComponent } from "../../theme/new-confirm-modal/confirm-modal.component";
+import { FieldModules } from "../custom-fields.constant";
 
 export interface CustomFieldsInterface {
     fieldName: string;
@@ -38,26 +41,46 @@ export class CustomFieldsListComponent implements OnInit, OnDestroy {
     };
     /** Holds get all custom fields api response */
     public customFieldsList: any = {};
+    /** Available field modules list */
+    public fieldModules: any[] = [];
+    /** True if translations are loaded */
+    public translationsLoaded: boolean = false;
     /** Observable to unsubscribe all the store listeners to avoid memory leaks */
     private destroyed$: ReplaySubject<boolean> = new ReplaySubject(1);
 
     constructor(
         private toasterService: ToasterService,
         private changeDetectorRef: ChangeDetectorRef,
-        private customFieldsService: CustomFieldsService
+        private customFieldsService: CustomFieldsService,
+        private dialog: MatDialog
     ) {
 
     }
 
+    /**
+     * Lifecycle hook for initialization
+     *
+     * @memberof CustomFieldsListComponent
+     */
     public ngOnInit(): void {
         this.getCustomFields();
     }
 
+    /**
+     * Lifecycle hook for destroy
+     *
+     * @memberof CustomFieldsListComponent
+     */
     public ngOnDestroy(): void {
         this.destroyed$.next(true);
         this.destroyed$.complete();
     }
 
+    /**
+     * Get list of custom fields
+     *
+     * @memberof CustomFieldsListComponent
+     */
     public getCustomFields(): void {
         this.customFieldsService.list(this.customFieldsRequest).pipe(takeUntil(this.destroyed$)).subscribe(response => {
             if (response) {
@@ -75,14 +98,33 @@ export class CustomFieldsListComponent implements OnInit, OnDestroy {
         });
     }
 
+    /**
+     * Delete custom field
+     *
+     * @param {*} customFieldUniqueName
+     * @memberof CustomFieldsListComponent
+     */
     public deleteCustomField(customFieldUniqueName: any): void {
-        this.customFieldsService.delete(customFieldUniqueName).pipe(takeUntil(this.destroyed$)).subscribe(response => {
-            //this.modalRef?.hide();
-            if (response?.status === "success") {
-                this.toasterService.showSnackBar("success", this.localeData?.custom_field_deleted);
-                this.getCustomFields();
-            } else {
-                this.toasterService.showSnackBar("error", response?.message);
+        let dialogRef = this.dialog.open(ConfirmModalComponent, {
+            width: '40%',
+            data: {
+                title: this.commonLocaleData?.app_delete,
+                body: this.localeData?.delete_custom_field_title,
+                ok: this.commonLocaleData?.app_yes,
+                cancel: this.commonLocaleData?.app_no
+            }
+        });
+
+        dialogRef.afterClosed().pipe(take(1)).subscribe(response => {
+            if (response) {
+                this.customFieldsService.delete(customFieldUniqueName).pipe(takeUntil(this.destroyed$)).subscribe(response => {
+                    if (response?.status === "success") {
+                        this.toasterService.showSnackBar("success", this.localeData?.custom_field_deleted);
+                        this.getCustomFields();
+                    } else {
+                        this.toasterService.showSnackBar("error", response?.message);
+                    }
+                });
             }
         });
     }
@@ -91,10 +133,28 @@ export class CustomFieldsListComponent implements OnInit, OnDestroy {
      * Page change event handler
      *
      * @param {PageChangedEvent} event Page changed event
-     * @memberof CustomFieldsComponent
+     * @memberof CustomFieldsListComponent 
      */
     public pageChanged(event: PageChangedEvent): void {
         this.customFieldsRequest.page = event.page;
         this.getCustomFields();
+    }
+
+    /**
+     * Callback for translation response complete
+     *
+     * @param {boolean} event
+     * @memberof CustomFieldsListComponent
+     */
+    public translationComplete(event: boolean): void {
+        if (event) {
+            this.fieldModules = [
+                { name: this.localeData?.modules?.account, uniqueName: FieldModules.Account },
+                { name: this.localeData?.modules?.stock, uniqueName: FieldModules.Stock },
+                { name: this.localeData?.modules?.group, uniqueName: FieldModules.Group }
+            ];
+            this.translationsLoaded = true;
+            this.changeDetectorRef.detectChanges();
+        }
     }
 }
